Fixed security hole in validateCredentials method that didn't gracefully fail on API auth response
The LSUCS Auth API returns an error object if you miss off the password. The validateCredentials method was assuming that the Auth API would fail with false on everything, which is not true. The Auth API only sends fail if the actual response to the API was correct but the credentials weren't valid.