public/controllers: new criteria for food order
User needs to have a ticket (any ticket) that is activated, and it needs to have a seat if not a visitor, as well as be during lan times.

The auth api has also been updated to properly disallow visitors form signing on to the LAN network as a result.